Extract PackageUpgradesIT from AbstractFuncIT
